+++ NEWS.md
@@ -1,3 +1,8 @@
+# rmcorr 0.7.0
+* Added geom for simplified graphing with ggplot2: geom_rmc (need to add unit tests)
+* Added vignette for overfitting/pseudoreplication (near final draft)
+* Updated roxygen to 7.3.2 and now using _PACKAGE
+
 # rmcorr 0.6.0
 
 * Added HCAHPS2022 dataset and description  
@@ -83,6 +88,3 @@
 * column names can be entered as strings and dynamically
 * dataset parameter is no longer required in the plot.rmc function
 * Added a `NEWS.md` file to track changes to the package.
-
-
-